refactor: defactoring deps usage analizer
