+++ dependent-usage-analyzer/tools/index.js
@@ -1,0 +1,4 @@
+const { analyzeProject } = require('./analyzeProject');
+const { findProjectsToAnalyze } = require('./findProjectsToAnalyze');
+
+module.exports = { analyzeProject, findProjectsToAnalyze };
