+++ dependent-usage-analyzer/utils/index.js
@@ -1,0 +1,5 @@
+const { getProjectFiles } = require('./getProjectFiles');
+const { getPackageInfo } = require('./getPackageInfo');
+const { getComponentUsagesInFiles } = require('./getComponentUsagesInFiles');
+
+module.exports = { getProjectFiles, getPackageInfo, getComponentUsagesInFiles };
